New IOrderDriverPort interface implementation
The `using` declarations and the `IOrderService` interface were removed from the `IOrderService.cs` file.
In the `OrdersController.cs` file, changed the implementation of the `OrdersController` class to implement the `IOrderDriverPort` interface instead of `IOrderService`.
A new `IOrderDriverPort.cs` file was added containing the necessary `using` declarations and the `IOrderDriverPort` interface definition, which includes the `CreateOrder` and `GetOrderById` methods.